Add some modals, improve modal UX
