+++ client/components/copy-list.vue
@@ -3,20 +3,22 @@
 </style>
 
 <template>
-    <div class="lpDialog" id="copyListDialog" v-if="shown">
-        <h2>Choose the list to copy</h2>
-        <select id="listToCopy" v-model="listId">
-            <option v-for="list in library.lists" :value="list.id">{{list.name}}</option>
-        </select>
-        <br /><br />
-        <p class="warning"><b>Note:</b> Copying a list will link the items between your lists. Updating an item in one list will alter that item in all other lists that item is in.</p>
-        <a v-on:click="copyList" class="lpButton" id="copyConfirm">Copy List</a>
-        <a v-on:click="closeModal" class="lpButton close">Cancel</a>
+    <div v-if="shown">
+        <div class="lpDialog" id="copyListDialog">
+            <h2>Choose the list to copy</h2>
+            <select id="listToCopy" v-model="listId">
+                <option v-for="list in library.lists" :value="list.id">{{list.name}}</option>
+            </select>
+            <br /><br />
+            <p class="warning"><b>Note:</b> Copying a list will link the items between your lists. Updating an item in one list will alter that item in all other lists that item is in.</p>
+            <a v-on:click="copyList" class="lpButton" id="copyConfirm">Copy List</a>
+            <a v-on:click="closeModal" class="lpButton close">Cancel</a>
+        </div>
+        <div v-on:click="closeModal" class="lpModalOverlay"></div>
     </div>
 </template>
 
 <script>
-
 const modalMixin = require("../mixins/modal-mixin.js");
 
 export default {
@@ -43,7 +45,7 @@ export default {
     },
     beforeMount: function() {
         bus.$on("copyList", () => {
-            this.shown = true;
+            this.openModal();
         });
     }
 }
