--- client/components/modal-blackout.vue
@@ -1,16 +1,0 @@
-<style lang="scss">
-#dashboard .hello-world {
-    font-size: 15px;
-    line-height: 1.8;
-}
-</style>
-
-<template>
-    <div class="lpModalOverlay lpBlackout"></div>
-</template>
-
-<script>
-export default {
-    name: "modalBlackout"
-}
-</script>
