+++ client/mixins/modal-mixin.js
@@ -7,6 +7,22 @@ var modalMixin = {
     methods: {
         closeModal: function() {
             this.shown = false;
+            this.unbindEscapeToClose();
+        },
+        openModal: function() {
+            this.shown = true;
+            this.bindEscapeToClose();
+        },
+        bindEscapeToClose: function() {
+            window.addEventListener('keyup', this.closeOnEscape);
+        },
+        unbindEscapeToClose: function() {
+            window.removeEventListener('keyup', this.closeOnEscape);
+        },
+        closeOnEscape: function(evt) {
+            if (evt.keyCode === 27) {
+                this.closeModal();
+            }
         }
     },
     beforeMount: function() {
